+++ src/otx/algo/common/layers/transformer_layers.py
@@ -71,9 +71,9 @@ class TransformerEncoderLayer(nn.Module):
             src = self.norm1(src)
         q = k = self.with_pos_embed(src, pos_embed)
         if self.key_mask:
-            src = self.self_attn(q, k, value=src, key_padding_mask=src_mask)[0]
+            src, _ = self.self_attn(q, k, value=src, key_padding_mask=src_mask, need_weights=False)
         else:
-            src, _ = self.self_attn(q, k, value=src, attn_mask=src_mask)
+            src, _ = self.self_attn(q, k, value=src, attn_mask=src_mask, need_weights=False)
 
         src = residual + self.dropout1(src)
         if not self.normalize_before: