+++ src/otx/tools/converter.py
@@ -338,9 +338,10 @@ class ConfigConverter:
                 config["callbacks"].pop(idx)
 
         def update_early_stop_patience(param_value: int) -> None:
-            for callback in config["callbacks"]:
-                if callback["class_path"] == "otx.algo.callbacks.adaptive_early_stopping.EarlyStoppingWithWarmup":
-                    callback["init_args"]["patience"] = param_value
+            if "callbacks" in config and config["callbacks"] is not None:
+                for callback in config["callbacks"]:
+                    if callback["class_path"] == "otx.algo.callbacks.adaptive_early_stopping.EarlyStoppingWithWarmup":
+                        callback["init_args"]["patience"] = param_value
                     break
 
         def update_use_adaptive_interval(param_value: bool) -> None:
@@ -491,14 +492,16 @@ class ConfigConverter:
             fail_untyped=False,
         )
         # Update callbacks & logger dir as engine.work_dir
-        for callback in config["callbacks"]:
-            if "init_args" in callback and "dirpath" in callback["init_args"]:
-                callback["init_args"]["dirpath"] = engine.work_dir
-        for logger in config["logger"]:
-            if "save_dir" in logger["init_args"]:
-                logger["init_args"]["save_dir"] = engine.work_dir
-            if "log_dir" in logger["init_args"]:
-                logger["init_args"]["log_dir"] = engine.work_dir
+        if "callbacks" in config and config["callbacks"] is not None:
+            for callback in config["callbacks"]:
+                if "init_args" in callback and "dirpath" in callback["init_args"]:
+                    callback["init_args"]["dirpath"] = engine.work_dir
+        if "logger" in config and config["logger"] is not None:
+            for logger in config["logger"]:
+                if "save_dir" in logger["init_args"]:
+                    logger["init_args"]["save_dir"] = engine.work_dir
+                if "log_dir" in logger["init_args"]:
+                    logger["init_args"]["log_dir"] = engine.work_dir
         instantiated_kwargs = engine_parser.instantiate_classes(Namespace(**config))
 
         train_kwargs = {k: v for k, v in instantiated_kwargs.items() if k in train_arguments}