+++ tests/unit/core/types/test_export.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2024 Intel Corporation
+# Copyright (C) 2024-2025 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 from copy import deepcopy
@@ -13,6 +13,7 @@ from otx.core.types.export import TaskLevelExportParameters
 def test_wrap(fxt_label_info, task_type):
     params = TaskLevelExportParameters(
         model_type="dummy model",
+        model_name="dummy model name",
         task_type=task_type,
         label_info=fxt_label_info,
         optimization_config={},
@@ -55,7 +56,10 @@ def test_wrap(fxt_label_info, task_type):
     assert ("model_info", "tile_size") in metadata
     assert ("model_info", "tiles_overlap") in metadata
     assert ("model_info", "max_pred_number") in metadata
+
+    # misc
     assert ("model_info", "otx_version") in metadata
+    assert ("model_info", "model_name") in metadata
 
 
 def test_to_metadata_label_consistency(fxt_label_info):
@@ -64,6 +68,7 @@ def test_to_metadata_label_consistency(fxt_label_info):
 
     params = TaskLevelExportParameters(
         model_type="dummy model",
+        model_name="dummy model name",
         task_type="instance_segmentation",
         label_info=label_info,
         optimization_config={},